Remove unused code, Move file to Signup Folder
