+++ src/sections/About.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styles from '../styles/About.module.css';
 
 function About() {
