+++ src/sections/Contact.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styles from '../styles/Contact.module.css';
 import TextField from '@mui/material/TextField';
