+++ src/sections/Schedule.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styles from '../styles/Schedule.module.css';
 
 function Schedule() {
