+++ src/sections/SignupForm/AccountForm.tsx
@@ -1,4 +1,4 @@
-import { FormWrapper } from "../../FormWrapper";
+import { FormWrapper } from "./FormWrapper";
 
 type AccountData = {
   goal: string;
