+++ src/sections/SignupForm/AddressForm.tsx
@@ -1,4 +1,4 @@
-import { FormWrapper } from "../../FormWrapper";
+import { FormWrapper } from "./FormWrapper";
 
 type AddressData = {
   firstHack: string;
