+++ src/sections/SignupForm/UserForm.tsx
@@ -1,4 +1,4 @@
-import { FormWrapper } from "../../FormWrapper";
+import { FormWrapper } from "./FormWrapper";
 
 type UserData = {
   firstName: string;
