+++ src/sections/Themes.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import styles from '../styles/Themes.module.css';
 import themeImage from '../assets/hackthebreak_logo.png';
